code/requirements.txt
absl-py==0.11.0
advertools==0.11.1
aiodns==2.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.4.1
altair==4.1.0
altgraph==0.17
argon2-cffi==20.1.0
astetik==1.9.9
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.3
bert-for-tf2==0.13.5
bleach==3.2.1
blinker==1.4
blis==0.7.4
bokeh==2.2.3
boto==2.49.0
boto3==1.12.15
botocore==1.15.15
cachetools==4.1.0
catalogue==2.0.4
cchardet==2.1.7
certifi==2019.11.28
cffi==1.14.0
chances==0.1.9
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.1
constantly==15.1.0
contextualSpellCheck==0.4.1
contextvars==2.4
cryptography==2.8
cssselect==1.1.0
curl==0.0.1
cycler==0.10.0
cymem==2.0.5
dask==2021.1.1
dask-glm==0.2.0
dask-ml==1.9.0
dataclasses==0.7
datasets==1.3.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distributed==2021.1.1
docker-pycreds==0.4.0
docutils==0.15.2
editdistance==0.5.3
elasticsearch==7.12.1
eli5==0.10.1
emoji==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
enum-compat==0.0.3
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.0.0/es_core_news_md-3.0.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0-py3-none-any.whl
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.0.0/es_dep_news_trf-3.0.0-py3-none-any.whl
et-xmlfile==1.0.1
fake-useragent==0.1.11
fasttext==0.9.2
filelock==3.0.12
flatbuffers==1.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
gensim==3.8.1
geographiclib==1.50
geonamescache==1.1.0
geopy==2.1.0
gitdb==4.0.5
GitPython==3.1.11
glove-python==0.1.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
googletransx==2.4.2
graphviz==0.13.2
grpcio==1.32.0
gviz-api==1.9.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
huggingface-hub==0.0.19
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==1.6.1
incremental==21.3.0
install==1.3.1
ipykernel==5.3.4
ipynb-py-convert==0.4.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kashgari==1.1.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-attention==1.0.0
keras-bert==0.81.0
keras-embed-sim==0.7.0
keras-gpt-2==0.14.0
keras-hypetune==0.1.1
keras-layer-normalization==0.14.0
keras-multi-head==0.22.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.41.0
keras-transformer==0.32.0
kerasplotlib==0.1.6
kiwisolver==1.1.0
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mlxtend==0.18.0
msgpack==1.0.2
multidict==5.1.0
multipledispatch==0.6.0
multiprocess==0.70.10
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
nlp==0.3.0
nltk==3.4.5
notebook==6.1.4
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.4
pandocfilters==1.4.3
params-flow==0.7.4
parsel==1.6.0
parso==0.7.0
partd==1.1.0
pathtools==0.1.2
pathy==0.6.0
patsy==0.5.1
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.1.0
pipelinehelper==0.7.7
pkbar==0.4
pkg_resources==0.0.0
plac==1.1.3
plotly==4.14.1
preshed==3.0.5
priority==1.3.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.12.2
psutil==5.7.3
ptyprocess==0.6.0
py-params==0.8.3
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycares==3.1.1
pycparser==2.19
pydantic==1.7.4
pydeck==0.5.0
PyDispatcher==2.0.5
pydot==1.4.1
Pygments==2.6.1
PyInstaller==4.0.dev0+a1f92c6a08
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.17.3
PySocks==1.7.1
pyspellchecker==0.6.2
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.2
qtconsole==5.1.1
QtPy==1.9.0
queuelib==1.6.1
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.38
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.5.4
Scrapy==2.5.0
seaborn==0.11.0
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.91
sentry-sdk==0.19.1
seqeval==0.0.10
service-identity==21.1.0
shap==0.39.0
shortuuid==1.0.1
simpletransformers==0.48.14
six==1.15.0
sklearn==0.0
slicer==0.0.7
smart-open==5.1.0
smmap==3.0.4
sortedcontainers==2.3.0
soupsieve==2.2.1
spacy==3.0.6
spacy-alignments==0.8.3
spacy-legacy==3.0.6
spacy-transformers==1.0.3
srsly==2.4.1
stanza==1.2.1
statsmodels==0.11.1
streamlit==0.69.2
subprocess32==3.5.4
symspellpy==6.7.0
tabulate==0.8.7
talos==1.0
tblib==1.7.0
tensorboard==2.4.1
tensorboard-plugin-profile==2.4.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==3.1.0
tensorflow-estimator==2.4.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.22.2
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
thinc==8.0.7
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.1
toolz==0.11.1
torch==1.7.1
torchdata==0.2.0
torchvision==0.6.1+cpu
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
transformers==4.11.3
tweet-preprocessor==0.6.0
twint @ file:///home/rafa_pepe/twint
Twisted==21.2.0
twython==3.8.2
typeguard==2.12.0
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.8
validators==0.18.1
w3lib==1.22.0
wandb==0.10.8
wasabi==0.8.2
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrangle==0.6.7
wrapt==1.12.1
xxhash==2.0.0
yarl==1.6.3
zict==2.0.0
zipp==3.1.0
zope.interface==5.4.0
